style: 회원가입 페이지1 컴포넌트 구현 완료(파일 위치 옮김)
